refactor: move functions to utils file
